src/main/java/task6_2D/OnTrackEnhanced.java
package task6_2D;

import java.util.HashMap;
import java.util.Map;

public class OnTrackEnhanced {
	
	private static final Map<String, String> students = new HashMap<>();
    private static final Map<String, String> tasks = new HashMap<>();
    private static final Map<String, String> courses = new HashMap<>();
    private static final Map<String, String> tutors = new HashMap<>();

    static {
        // Simulate some data
        students.put("12345", "John Doe");
        tasks.put("task123", "Sample Task");
        courses.put("courseABC", "Sample Course");
        tutors.put("tutor678", "Jane Smith");
    }

    // Advanced Progress Tracking
    public String getDetailedAnalytics(String studentId) {
        if (studentId == null || studentId.isEmpty()) {
            return "Invalid student ID.";
        }
        if (!students.containsKey(studentId)) {
            return "Student not found.";
        }
        // Simulate fetching detailed progress metrics
        return "Detailed progress metrics for student " + studentId + ":\n- Assignment performance\n- Time spent on tasks\n- Comparison with peers";
    }

    public String sendAutomatedReminders(String studentId) {
        if (studentId == null || studentId.isEmpty()) {
            return "Invalid student ID.";
        }
        if (!students.containsKey(studentId)) {
            return "Student not found.";
        }
        // Simulate sending reminders
        return "Automated reminders sent to student " + studentId + " for upcoming deadlines and pending tasks.";
    }

    // Improved Task Management
    public String createSubtask(String taskId, String subtaskDetails) {
        if (taskId == null || taskId.isEmpty() || subtaskDetails == null || subtaskDetails.isEmpty()) {
            return "Invalid parameters for subtask creation.";
        }
        if (!tasks.containsKey(taskId)) {
            return "Task not found.";
        }
        // Simulate subtask creation
        return "Subtask created for task " + taskId + " with details: " + subtaskDetails;
    }

    public String createCollaborativeWorkspace(String groupId, String documentDetails) {
        if (groupId == null || groupId.isEmpty() || documentDetails == null || documentDetails.isEmpty()) {
            return "Invalid parameters for collaborative workspace creation.";
        }
     // Simulate checking if the group exists
        boolean groupExists = checkIfGroupExists(groupId);
        
        if (groupExists) {
            // Simulate collaborative workspace creation
            return "Collaborative workspace created for group " + groupId + " with document details: " + documentDetails;
        } else {
            return "Group not found.";
        }
    }
    
    // Simulated method to check if the group exists
    private boolean checkIfGroupExists(String groupId) {
        // Simulate checking if the group exists in the system
        // For demonstration purposes, we'll assume that the group exists if the group ID starts with "group"
        return groupId.startsWith("group");
    }

    // Submission Flexibility
    public String submitTaskInMultipleFormats(String taskId, String fileFormat, String fileDetails) {
        if (taskId == null || taskId.isEmpty() || fileFormat == null || fileFormat.isEmpty() || fileDetails == null || fileDetails.isEmpty()) {
            return "Invalid parameters for task submission.";
        }
        if (!tasks.containsKey(taskId)) {
            return "Task not found.";
        }
        // Simulate task submission
        return "Task " + taskId + " submitted in format " + fileFormat + " with file details: " + fileDetails;
    }

    public String saveDraftSubmission(String taskId, String draftDetails) {
        if (taskId == null || taskId.isEmpty() || draftDetails == null || draftDetails.isEmpty()) {
            return "Invalid parameters for draft submission.";
        }
        if (!tasks.containsKey(taskId)) {
            return "Task not found.";
        }
        // Simulate draft saving
        return "Draft saved for task " + taskId + " with draft details: " + draftDetails;
    }

    // Enhanced Communication Tools
    public String startVideoConference(String studentId, String tutorId) {
        if (studentId == null || studentId.isEmpty() || tutorId == null || tutorId.isEmpty()) {
            return "Invalid parameters for video conference.";
        }
        if (!students.containsKey(studentId) || !tutors.containsKey(tutorId)) {
            return "Student or tutor not found.";
        }
        // Simulate starting a video conference
        return "Video conference started between student " + studentId + " and tutor " + tutorId;
    }

    public String createDiscussionForum(String courseId, String forumDetails) {
        if (courseId == null || courseId.isEmpty() || forumDetails == null || forumDetails.isEmpty()) {
            return "Invalid parameters for discussion forum creation.";
        }
        if (!courses.containsKey(courseId)) {
            return "Course not found.";
        }
        // Simulate forum creation
        return "Discussion forum created for course " + courseId + " with forum details: " + forumDetails;
    }

    // Detailed Feedback and Grading
    public String provideDetailedFeedback(String taskId, String feedbackDetails) {
        if (taskId == null || taskId.isEmpty() || feedbackDetails == null || feedbackDetails.isEmpty()) {
            return "Invalid parameters for feedback.";
        }
        if (!tasks.containsKey(taskId)) {
            return "Task not found.";
        }
        // Simulate providing feedback
        return "Feedback provided for task " + taskId + " with feedback details: " + feedbackDetails;
    }

    public String provideRubricBasedAssessment(String taskId, String rubricDetails) {
        if (taskId == null || taskId.isEmpty() || rubricDetails == null || rubricDetails.isEmpty()) {
            return "Invalid parameters for rubric-based assessment.";
        }
        if (!tasks.containsKey(taskId)) {
            return "Task not found.";
        }
        // Simulate rubric-based assessment
        return "Rubric-based assessment provided for task " + taskId + " with rubric details: " + rubricDetails;
    }

    // Automated Extension Requests
    public String autoApproveExtensionRequest(String taskId, String extensionCriteria) {
        if (taskId == null || taskId.isEmpty() || extensionCriteria == null || extensionCriteria.isEmpty()) {
            return "Invalid parameters for extension request.";
        }
        if (!tasks.containsKey(taskId)) {
            return "Task not found.";
        }
        // Simulate auto-approval of extension
        return "Extension request auto-approved for task " + taskId + " based on criteria: " + extensionCriteria;
    }

    // Tutor Functions
    // Advanced Assignment Review
    public String detectPlagiarism(String submissionDetails) {
        if (submissionDetails == null || submissionDetails.isEmpty()) {
            return "Invalid parameters for plagiarism detection.";
        }
        // Simulate plagiarism detection
        return "Plagiarism detection completed for submission with details: " + submissionDetails;
    }

    // Enhanced Communication Tools
    public String setVirtualOfficeHours(String tutorId, String scheduleDetails) {
        if (tutorId == null || tutorId.isEmpty() || scheduleDetails == null || scheduleDetails.isEmpty()) {
            return "Invalid parameters for setting virtual office hours.";
        }
        if (!tutors.containsKey(tutorId)) {
            return "Tutor not found.";
        }
        // Simulate setting virtual office hours
        return "Virtual office hours set for tutor " + tutorId + " with schedule details: " + scheduleDetails;
    }

    public String useAutomatedResponseTemplate(String templateId, String responseDetails) {
        if (templateId == null || templateId.isEmpty() || responseDetails == null || responseDetails.isEmpty()) {
            return "Invalid parameters for automated response.";
        }
        // Simulate using automated response template
        return "Automated response sent using template " + templateId + " with response details: " + responseDetails;
    }

    // Performance Analytics
    public String generateStudentProgressReport(String studentId) {
        if (studentId == null || studentId.isEmpty()) {
            return "Invalid student ID.";
        }
        if (!students.containsKey(studentId)) {
            return "Student not found.";
        }
        // Simulate generating student progress report
        return "Progress report generated for student " + studentId + ": \n- Overall progress\n- Strengths\n- Areas for improvement";
    }

    public String generateCourseAnalytics(String courseId) {
        if (courseId == null || courseId.isEmpty()) {
            return "Invalid course ID.";
        }
        if (!courses.containsKey(courseId)) {
            return "Course not found.";
        }
        // Simulate generating course analytics
        return "Analytics generated for course " + courseId + ": \n- Completion rates\n- Average grades\n- Student engagement levels";
    }

    // Resource Sharing
    public String createInteractiveContent(String courseId, String contentDetails) {
        if (courseId == null || courseId.isEmpty() || contentDetails == null || contentDetails.isEmpty()) {
            return "Invalid parameters for interactive content creation.";
        }
        if (!courses.containsKey(courseId)) {
            return "Course not found.";
        }
        // Simulate creating interactive content
        return "Interactive content created for course " + courseId + " with details: " + contentDetails;
    }

    // Workflow Optimization
    public String automateTask(String taskId, String automationDetails) {
        if (taskId == null || taskId.isEmpty() || automationDetails == null || automationDetails.isEmpty()) {
            return "Invalid parameters for task automation.";
        }
        if (!tasks.containsKey(taskId)) {
            return "Task not found.";
        }
        // Simulate task automation
        return "Task " + taskId + " automated with details: " + automationDetails;
    }

    public String coordinateWithTutors(String tutorId, String coordinationDetails) {
        if (tutorId == null || tutorId.isEmpty() || coordinationDetails == null || coordinationDetails.isEmpty()) {
            return "Invalid parameters for tutor coordination.";
        }
        if (!tutors.containsKey(tutorId)) {
            return "Tutor not found.";
        }
        // Simulate coordinating with other tutors
        return "Coordination completed with tutor " + tutorId + " with details: " + coordinationDetails;
    }

    // Feedback Management
    public String collectStudentFeedback(String courseId, String feedbackDetails) {
        if (courseId == null || courseId.isEmpty() || feedbackDetails == null || feedbackDetails.isEmpty()) {
            return "Invalid parameters for feedback collection.";
        }
        if (!courses.containsKey(courseId)) {
            return "Course not found.";
        }
        // Simulate collecting student feedback
        return "Feedback collected for course " + courseId + " with details: " + feedbackDetails;
    }

}
