src/test/java/task6_2D/OnTrackEnhancedTest.java
package task6_2D;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotEquals;
import static org.junit.Assert.assertTrue;

import org.junit.Test;

public class OnTrackEnhancedTest {

	private final OnTrackEnhanced onTrackEnhanced = new OnTrackEnhanced();

	// Advanced Progress Tracking
    @Test
    public void testGetDetailedAnalytics() {
        // Right
        assertEquals("Detailed progress metrics for student 12345:\n- Assignment performance\n- Time spent on tasks\n- Comparison with peers", onTrackEnhanced.getDetailedAnalytics("12345"));
        // Boundary (non-existing student)
        assertEquals("Student not found.", onTrackEnhanced.getDetailedAnalytics("99999"));
        // Invalid input
        assertEquals("Invalid student ID.", onTrackEnhanced.getDetailedAnalytics(null));
        assertEquals("Invalid student ID.", onTrackEnhanced.getDetailedAnalytics(""));
        // Cross-check (make sure no analytics for invalid student)
        assertEquals("Student not found.", onTrackEnhanced.getDetailedAnalytics("00000"));
        // Inverse (check for empty result)
        assertNotEquals("Detailed progress metrics for student 12345:\n- Assignment performance\n- Time spent on tasks\n- Comparison with peers", onTrackEnhanced.getDetailedAnalytics("67890"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.getDetailedAnalytics("12345");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    @Test
    public void testSendAutomatedReminders() {
        // Right
        assertEquals("Automated reminders sent to student 12345 for upcoming deadlines and pending tasks.", onTrackEnhanced.sendAutomatedReminders("12345"));
        // Boundary (non-existing student)
        assertEquals("Student not found.", onTrackEnhanced.sendAutomatedReminders("99999"));
        // Invalid input
        assertEquals("Invalid student ID.", onTrackEnhanced.sendAutomatedReminders(null));
        assertEquals("Invalid student ID.", onTrackEnhanced.sendAutomatedReminders(""));
        // Cross-check (ensure reminders not sent for invalid student)
        assertEquals("Student not found.", onTrackEnhanced.sendAutomatedReminders("00000"));
        // Inverse (check for different student)
        assertNotEquals("Automated reminders sent to student 12345 for upcoming deadlines and pending tasks.", onTrackEnhanced.sendAutomatedReminders("67890"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.sendAutomatedReminders("12345");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    // Improved Task Management
    @Test
    public void testCreateSubtask() {
        // Right
        assertEquals("Subtask created for task task123 with details: Sample subtask", onTrackEnhanced.createSubtask("task123", "Sample subtask"));
        // Boundary (non-existing task)
        assertEquals("Task not found.", onTrackEnhanced.createSubtask("task999", "Sample subtask"));
        // Invalid input
        assertEquals("Invalid parameters for subtask creation.", onTrackEnhanced.createSubtask(null, "Sample subtask"));
        assertEquals("Invalid parameters for subtask creation.", onTrackEnhanced.createSubtask("task123", null));
        assertEquals("Invalid parameters for subtask creation.", onTrackEnhanced.createSubtask("", "Sample subtask"));
        assertEquals("Invalid parameters for subtask creation.", onTrackEnhanced.createSubtask("task123", ""));
        // Cross-check (ensure subtasks not created for invalid task)
        assertEquals("Task not found.", onTrackEnhanced.createSubtask("task000", "Sample subtask"));
        // Inverse (check for another task)
        assertNotEquals("Subtask created for task task123 with details: Sample subtask", onTrackEnhanced.createSubtask("task456", "Sample subtask"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.createSubtask("task123", "Sample subtask");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    @Test
    public void testCreateCollaborativeWorkspace() {
        // Right
        assertEquals("Collaborative workspace created for group group123 with document details: Shared document", onTrackEnhanced.createCollaborativeWorkspace("group123", "Shared document"));
        // Boundary (non-existing group)
        assertEquals("Group not found.", onTrackEnhanced.createCollaborativeWorkspace("grou999", "Shared document"));
        // Invalid input
        assertEquals("Invalid parameters for collaborative workspace creation.", onTrackEnhanced.createCollaborativeWorkspace(null, "Shared document"));
        assertEquals("Invalid parameters for collaborative workspace creation.", onTrackEnhanced.createCollaborativeWorkspace("group123", null));
        assertEquals("Invalid parameters for collaborative workspace creation.", onTrackEnhanced.createCollaborativeWorkspace("", "Shared document"));
        assertEquals("Invalid parameters for collaborative workspace creation.", onTrackEnhanced.createCollaborativeWorkspace("group123", ""));
        // Cross-check (ensure workspace not created for invalid group)
        assertEquals("Group not found.", onTrackEnhanced.createCollaborativeWorkspace("grou000", "Shared document"));
        // Inverse (check for another group)
        assertNotEquals("Collaborative workspace created for group group123 with document details: Shared document", onTrackEnhanced.createCollaborativeWorkspace("group456", "Shared document"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.createCollaborativeWorkspace("group123", "Shared document");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    // Submission Flexibility
    @Test
    public void testSubmitTaskInMultipleFormats() {
        // Right
        assertEquals("Task task123 submitted in format PDF with file details: Sample file", onTrackEnhanced.submitTaskInMultipleFormats("task123", "PDF", "Sample file"));
        // Boundary (non-existing task)
        assertEquals("Task not found.", onTrackEnhanced.submitTaskInMultipleFormats("task999", "PDF", "Sample file"));
        // Invalid input
        assertEquals("Invalid parameters for task submission.", onTrackEnhanced.submitTaskInMultipleFormats(null, "PDF", "Sample file"));
        assertEquals("Invalid parameters for task submission.", onTrackEnhanced.submitTaskInMultipleFormats("task123", null, "Sample file"));
        assertEquals("Invalid parameters for task submission.", onTrackEnhanced.submitTaskInMultipleFormats("task123", "PDF", null));
        assertEquals("Invalid parameters for task submission.", onTrackEnhanced.submitTaskInMultipleFormats("", "PDF", "Sample file"));
        assertEquals("Invalid parameters for task submission.", onTrackEnhanced.submitTaskInMultipleFormats("task123", "", "Sample file"));
        assertEquals("Invalid parameters for task submission.", onTrackEnhanced.submitTaskInMultipleFormats("task123", "PDF", ""));
        // Cross-check (ensure submission not accepted for invalid task)
        assertEquals("Task not found.", onTrackEnhanced.submitTaskInMultipleFormats("task000", "PDF", "Sample file"));
        // Inverse (check for different format)
        assertNotEquals("Task task123 submitted in format PDF with file details: Sample file", onTrackEnhanced.submitTaskInMultipleFormats("task123", "DOCX", "Sample file"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.submitTaskInMultipleFormats("task123", "PDF", "Sample file");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    @Test
    public void testSaveDraftSubmission() {
        // Right
        assertEquals("Draft saved for task task123 with draft details: Draft content", onTrackEnhanced.saveDraftSubmission("task123", "Draft content"));
        // Boundary (non-existing task)
        assertEquals("Task not found.", onTrackEnhanced.saveDraftSubmission("task999", "Draft content"));
        // Invalid input
        assertEquals("Invalid parameters for draft submission.", onTrackEnhanced.saveDraftSubmission(null, "Draft content"));
        assertEquals("Invalid parameters for draft submission.", onTrackEnhanced.saveDraftSubmission("task123", null));
        assertEquals("Invalid parameters for draft submission.", onTrackEnhanced.saveDraftSubmission("", "Draft content"));
        assertEquals("Invalid parameters for draft submission.", onTrackEnhanced.saveDraftSubmission("task123", ""));
        // Cross-check (ensure draft not saved for invalid task)
        assertEquals("Task not found.", onTrackEnhanced.saveDraftSubmission("task000", "Draft content"));
        // Inverse (check for different draft details)
        assertNotEquals("Draft saved for task task123 with draft details: Draft content", onTrackEnhanced.saveDraftSubmission("task123", "Different draft content"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.saveDraftSubmission("task123", "Draft content");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    // Enhanced Communication Tools
    @Test
    public void testStartVideoConference() {
        // Right
        assertEquals("Video conference started between student 12345 and tutor tutor678", onTrackEnhanced.startVideoConference("12345", "tutor678"));
        // Boundary (non-existing student or tutor)
        assertEquals("Student or tutor not found.", onTrackEnhanced.startVideoConference("99999", "tutor678"));
        assertEquals("Student or tutor not found.", onTrackEnhanced.startVideoConference("12345", "99999"));
        // Invalid input
        assertEquals("Invalid parameters for video conference.", onTrackEnhanced.startVideoConference(null, "tutor678"));
        assertEquals("Invalid parameters for video conference.", onTrackEnhanced.startVideoConference("12345", null));
        assertEquals("Invalid parameters for video conference.", onTrackEnhanced.startVideoConference("", "tutor678"));
        assertEquals("Invalid parameters for video conference.", onTrackEnhanced.startVideoConference("12345", ""));
        // Cross-check (ensure conference not started for invalid participants)
        assertEquals("Student or tutor not found.", onTrackEnhanced.startVideoConference("00000", "00000"));
        // Inverse (check for different participants)
        assertNotEquals("Video conference started between student 12345 and tutor tutor679", onTrackEnhanced.startVideoConference("12345", "tutor678"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.startVideoConference("12345", "67890");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    @Test
    public void testCreateDiscussionForum() {
        // Right
        assertEquals("Discussion forum created for course courseABC with forum details: Forum details", onTrackEnhanced.createDiscussionForum("courseABC", "Forum details"));
        // Boundary (non-existing course)
        assertEquals("Course not found.", onTrackEnhanced.createDiscussionForum("course999", "Forum details"));
        // Invalid input
        assertEquals("Invalid parameters for discussion forum creation.", onTrackEnhanced.createDiscussionForum(null, "Forum details"));
        assertEquals("Invalid parameters for discussion forum creation.", onTrackEnhanced.createDiscussionForum("courseABC", null));
        assertEquals("Invalid parameters for discussion forum creation.", onTrackEnhanced.createDiscussionForum("", "Forum details"));
        assertEquals("Invalid parameters for discussion forum creation.", onTrackEnhanced.createDiscussionForum("courseABC", ""));
        // Cross-check (ensure forum not created for invalid course)
        assertEquals("Course not found.", onTrackEnhanced.createDiscussionForum("course000", "Forum details"));
        // Inverse (check for different forum details)
        assertNotEquals("Discussion forum created for course courseABC with details: Forum details", onTrackEnhanced.createDiscussionForum("courseABC", "Different forum details"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.createDiscussionForum("courseABC", "Forum details");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    // Detailed Feedback and Grading
    @Test
    public void testProvideDetailedFeedback() {
        // Right
        assertEquals("Feedback provided for task task123 with feedback details: Feedback content", onTrackEnhanced.provideDetailedFeedback("task123", "Feedback content"));
        // Boundary (non-existing task)
        assertEquals("Task not found.", onTrackEnhanced.provideDetailedFeedback("task999", "Feedback content"));
        // Invalid input
        assertEquals("Invalid parameters for feedback.", onTrackEnhanced.provideDetailedFeedback(null, "Feedback content"));
        assertEquals("Invalid parameters for feedback.", onTrackEnhanced.provideDetailedFeedback("task123", null));
        assertEquals("Invalid parameters for feedback.", onTrackEnhanced.provideDetailedFeedback("", "Feedback content"));
        assertEquals("Invalid parameters for feedback.", onTrackEnhanced.provideDetailedFeedback("task123", ""));
        // Cross-check (ensure feedback not provided for invalid task)
        assertEquals("Task not found.", onTrackEnhanced.provideDetailedFeedback("task000", "Feedback content"));
        // Inverse (check for different feedback details)
        assertNotEquals("Feedback provided for task task123 with details: Feedback content", onTrackEnhanced.provideDetailedFeedback("task123", "Different feedback content"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.provideDetailedFeedback("task123", "Feedback content");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    @Test
    public void testProvideRubricBasedAssessment() {
        // Right
        assertEquals("Rubric-based assessment provided for task task123 with rubric details: Rubric content", onTrackEnhanced.provideRubricBasedAssessment("task123", "Rubric content"));
        // Boundary (non-existing task)
        assertEquals("Task not found.", onTrackEnhanced.provideRubricBasedAssessment("task999", "Rubric content"));
        // Invalid input
        assertEquals("Invalid parameters for rubric-based assessment.", onTrackEnhanced.provideRubricBasedAssessment(null, "Rubric content"));
        assertEquals("Invalid parameters for rubric-based assessment.", onTrackEnhanced.provideRubricBasedAssessment("task123", null));
        assertEquals("Invalid parameters for rubric-based assessment.", onTrackEnhanced.provideRubricBasedAssessment("", "Rubric content"));
        assertEquals("Invalid parameters for rubric-based assessment.", onTrackEnhanced.provideRubricBasedAssessment("task123", ""));
        // Cross-check (ensure rubric not provided for invalid task)
        assertEquals("Task not found.", onTrackEnhanced.provideRubricBasedAssessment("task000", "Rubric content"));
        // Inverse (check for different rubric details)
        assertNotEquals("Rubric-based assessment provided for task task123 with details: Rubric content", onTrackEnhanced.provideRubricBasedAssessment("task123", "Different rubric content"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.provideRubricBasedAssessment("task123", "Rubric content");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    // Automated Extension Requests
    @Test
    public void testAutoApproveExtensionRequest() {
        // Right
        assertEquals("Extension request auto-approved for task task123 based on criteria: Criteria details", onTrackEnhanced.autoApproveExtensionRequest("task123", "Criteria details"));
        // Boundary (non-existing task)
        assertEquals("Task not found.", onTrackEnhanced.autoApproveExtensionRequest("task999", "Criteria details"));
        // Invalid input
        assertEquals("Invalid parameters for extension request.", onTrackEnhanced.autoApproveExtensionRequest(null, "Criteria details"));
        assertEquals("Invalid parameters for extension request.", onTrackEnhanced.autoApproveExtensionRequest("task123", null));
        assertEquals("Invalid parameters for extension request.", onTrackEnhanced.autoApproveExtensionRequest("", "Criteria details"));
        assertEquals("Invalid parameters for extension request.", onTrackEnhanced.autoApproveExtensionRequest("task123", ""));
        // Cross-check (ensure extension not auto-approved for invalid task)
        assertEquals("Task not found.", onTrackEnhanced.autoApproveExtensionRequest("task000", "Criteria details"));
        // Inverse (check for different criteria)
        assertNotEquals("Extension request auto-approved for task task123 with criteria: Criteria details", onTrackEnhanced.autoApproveExtensionRequest("task123", "Different criteria details"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.autoApproveExtensionRequest("task123", "Criteria details");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }

    // Tutor Functions
    // Advanced Assignment Review
    @Test
    public void testDetectPlagiarism() {
        // Right
        assertEquals("Plagiarism detection completed for submission with details: Submission content", onTrackEnhanced.detectPlagiarism("Submission content"));
        // Boundary (short submission)
        assertEquals("Plagiarism detection completed for submission with details: Short content", onTrackEnhanced.detectPlagiarism("Short content"));
        // Invalid input
        assertEquals("Invalid parameters for plagiarism detection.", onTrackEnhanced.detectPlagiarism(null));
        assertEquals("Invalid parameters for plagiarism detection.", onTrackEnhanced.detectPlagiarism(""));
        // Cross-check (ensure plagiarism not detected for invalid input)
        assertEquals("Invalid parameters for plagiarism detection.", onTrackEnhanced.detectPlagiarism(""));
        // Inverse (check for different content)
        assertNotEquals("Plagiarism detection completed for submission with details: Submission content", onTrackEnhanced.detectPlagiarism("Different content"));
        // Performance (quick response)
        long startTime = System.nanoTime();
        onTrackEnhanced.detectPlagiarism("Submission content");
        long endTime = System.nanoTime();
        assertTrue(endTime - startTime < 1_000_000); // Less than 1ms
    }
}
